add id to categories to fix renaming bug

+++ simplest_budget/app/src/main/java/clbisk/simplestbudget/ui/reusable/budgetcategories/modify/create/CreateBudgetCategoryForm.kt
@@ -12,7 +12,7 @@ import clbisk.simplestbudget.R
 @Composable
 fun CreateBudgetCategoryForm (
 	name: String,
-	limit: Long,
+	limit: Float,
 	isValid: Boolean,
 	onNameChange: (String) -> Unit,
 	onLimitChange: (String) -> Unit,
