+++ simplest_budget/app/src/main/java/clbisk/simplestbudget/ui/reusable/budgetcategories/modify/edit/EditBudgetCategoryForm.kt
@@ -12,7 +12,7 @@ import clbisk.simplestbudget.R
 @Composable
 fun EditBudgetCategoryForm (
 	name: String,
-	limit: Long,
+	limit: Float,
 	isLoaded: Boolean,
 	onNameInputChange: (String) -> Unit,
 	onLimitInputChange: (String) -> Unit,
