+++ simplest_budget/app/src/main/java/clbisk/simplestbudget/ui/reusable/model/CategoryState.kt
@@ -6,5 +6,5 @@ data class CategoryState(
 	val categoryName: String,
 	val loaded: Boolean = false,
 	val data: BudgetCategory? = null,
-	val transactionTotal: Long? = null,
+	val transactionTotal: Float? = null,
 )
